Refactoring add_global_variable
In order to make the method accept either a single value or an Array, an optional
parameter was chosen. This was used in order to avoid checking if pVariable is
an Array or not. It also must be said that the faltten method was used on
pVariable, since optional parameters in ruby are already an Array by default,
therefore, if the user pass an Array as parameter, without the flatten call,
pVariable would be an Array of an Array.

This commit closes #61

Signed-off-by: Lucas Moura <lucas.moura128@gmail.com>